+++ packages/fedify/src/runtime/mod.ts
@@ -1,0 +1,13 @@
+/**
+ * Runtime facilities for working with Activity Vocabulary objects.
+ *
+ * @deprecated This module has been moved to the `@fedify/vocab-runtime`
+ *             package.  Please use `@fedify/vocab-runtime` instead of
+ *             `@fedify/fedify/runtime`.  This module will be removed in
+ *             a future version.
+ *             See {@link https://github.com/fedify-dev/fedify/issues/444}
+ *             and {@link https://github.com/fedify-dev/fedify/pull/451}
+ *             for more details.
+ * @module
+ */
+export * from "@fedify/vocab-runtime";