+++ packages/vocab/src/mod.ts
@@ -55,3 +55,9 @@ export * from "./handle.ts";
 export * from "./lookup.ts";
 export * from "./type.ts";
 export * from "./vocab.ts";
+export { LanguageString } from "@fedify/vocab-runtime";
+export type {
+  DocumentLoader,
+  GetUserAgentOptions,
+  RemoteDocument,
+} from "@fedify/vocab-runtime";